feat: Added cy normal calculation function
