+++ src/ray_tracing/render_utils.c
@@ -41,3 +41,25 @@ t_rgb	get_colour(t_object *obj)
 	else
 		return (obj->obj.cylinder.colour);
 }
+
+t_coord	deduce_normal(t_coord p2p, t_coord normal)
+{
+	if (dot_prod(p2p, normal) < 0)
+		return(vect_mult(normal, -1));
+	return (normal);
+}
+
+t_coord	get_cy_normal(t_cy obj, t_coord intsct)
+{
+	double	dist;
+	t_coord	cross_res;
+	t_coord	p2p_vector;
+
+	p2p_vector = vect_subt(intsct, obj.center);
+	cross_res = cross_prod(obj.axis_vector, p2p_vector);
+	dist = vect_magnitude(cross_res, (t_coord){0, 0, 0, 1});
+	if (approx(dist, obj.radius) == 1) // The intsct point is on the curved surface
+		return (deduce_normal(normalize(p2p_vector), normalize(cross_res)));
+	else
+		return (deduce_normal(normalize(p2p_vector), obj.axis_vector));
+}
